ANDKR-1958 | 내주변 검색시 목록에서 필터 선택시 지역별 필터 삭제

+++ dailyHOTEL/app/src/main/java/com/daily/dailyhotel/screen/home/stay/inbound/filter/StayFilterActivity.java
@@ -42,7 +42,7 @@ public class StayFilterActivity extends BaseActivity<StayFilterPresenter>
     public static Intent newInstance(Context context, SearchStayResultTabPresenter.ListType listType//
         , String checkInDateTime, String checkOutDateTime//
         , DailyCategoryType categoryType, String viewType, StayFilter filter, StaySuggest suggest//
-        , ArrayList<String> categories, Location location, double radius, String searchWord)
+        , ArrayList<String> categories, Location location, float radius, String searchWord)
     {
         Intent intent = new Intent(context, StayFilterActivity.class);
         intent.putExtra(INTENT_EXTRA_DATA_LIST_TYPE, listType.name());
